Updated Master Exclamation Point

+++ src/main/java/com/foundgrounds/photos/PhotoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.foundgrounds.photos.Photo;
 
-public interface PhotoRepository extends JpaRepository<Photo, Long> {
+public interface PhotoRepository extends JpaRepository<Photo, Integer> {
 
 }
