+++ src/main/java/com/foundgrounds/users/UserRepository.java
@@ -2,8 +2,5 @@ package com.foundgrounds.users;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.foundgrounds.users.User;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
+public interface UserRepository extends JpaRepository<User, Integer> {
 }
